refactor(core,ui): 客观题并入 U/P 题库，is_objective 标记替代独立 O 型（#222 设计调整）
按用户决策重构（不再设立 O 题库）：
- schema：problems 新增 is_objective 布尔列，type CHECK 回 ('U','P')；
  0034 迁移（保留 0033 的新表与 runtime_config 可空）
- 权限：客观题套卷权限随 type（U 型 owner/admin、P 型仅 admin），
  无独立权限特例；答案/解析可见性同步（P 型仅 admin）
- API：客观题端点并入 /api/v1/problems/*（questions CRUD、submit、
  submissions），删除独立 /api/v1/objective 路由
- 前端：删除 /objective-papers 页面与导航入口；作答并入
  problems/[id].vue（ObjectiveAnswerForm 组件）、套卷编辑并入
  problems/[id]/edit.vue（ObjectivePaperEditor 组件）、创建页加
  「客观题套卷」开关（ProblemEditor objective prop）；题库列表在
  难度位置标记「客观题」；竞赛题面页改按 is_objective 判断
- 测试：40 个 objective 用例全绿（含新增 P 型套卷权限用例）；
  E2E 断言 is_objective 语义
- 文档：OpenSpec 主规范与归档 specs 同步 is_objective 设计，
  AGENTS.md 表清单更新
