+++ noj-core/drizzle/0034_known_rockslide.sql
@@ -1,0 +1,3 @@
+ALTER TABLE "problems" DROP CONSTRAINT "problems_type_check";--> statement-breakpoint
+ALTER TABLE "problems" ADD COLUMN "is_objective" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "problems" ADD CONSTRAINT "problems_type_check" CHECK ("problems"."type" IN ('U', 'P'));
