+++ noj-ui/pages/problems/new.vue
@@ -5,8 +5,11 @@ definePageMeta({
 
 const router = useRouter()
 
-function onSaved() {
-  router.replace("/my/problems")
+// 客观题套卷：创建 is_objective=true 的题目（type 可选 U/P，权限随类型）
+const isObjective = ref(false)
+
+function onSaved(id: string) {
+  router.replace(isObjective.value ? `/problems/${id}/edit` : "/my/problems")
 }
 </script>
 
@@ -14,6 +17,18 @@ function onSaved() {
   <div class="px-4 py-5 sm:px-7 sm:py-8 max-w-[860px] mx-auto">
     <h1 class="text-2xl font-bold text-text mb-5">创建题目</h1>
 
-    <ProblemEditor mode="create" initial-type="U" @saved="onSaved" />
+    <div class="mb-4 flex items-center gap-2 rounded-xl border border-border bg-white px-4 py-3">
+      <input
+        id="objective-toggle"
+        v-model="isObjective"
+        type="checkbox"
+        class="accent-primary"
+      />
+      <label for="objective-toggle" class="text-sm text-text">
+        客观题套卷<span class="text-text-muted">（单选/多选/判断小题，服务端即时判定，无需评测配置）</span>
+      </label>
+    </div>
+
+    <ProblemEditor mode="create" initial-type="U" :objective="isObjective" @saved="onSaved" />
   </div>
 </template>
