+++ openspec/changes/archive/2026-08-07-objective-questions/specs/contest-participation/spec.md
@@ -2,7 +2,7 @@
 
 ### Requirement: 竞赛客观题提交
 
-系统 SHALL 支持参赛者在竞赛中提交客观题套卷答案，使用 `POST /api/v1/objective/papers/:id/submit` 并携带 `contest_id`。提交前 SHALL 校验：竞赛存在且状态为 `running`、用户已注册参赛、套卷属于该竞赛的 contest_problems、该用户对该套卷在该竞赛下无既有提交（只允许一次）。
+系统 SHALL 支持参赛者在竞赛中提交客观题套卷答案，使用 `POST /api/v1/problems/:id/submit` 并携带 `contest_id`。提交前 SHALL 校验：竞赛存在且状态为 `running`、用户已注册参赛、套卷属于该竞赛的 contest_problems、该用户对该套卷在该竞赛下无既有提交（只允许一次）。
 
 #### Scenario: 竞赛期间提交套卷答案
 - **WHEN** 已注册参赛者在 running 竞赛中提交该竞赛题单内套卷的答案（首次）
